% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{check_table}
\alias{check_table}
\title{Table content check based on json schema template for uploaded user csv tables}
\usage{
check_table(
  user_table_csv_paths,
  table_names,
  json_template_path,
  delimiter = ","
)
}
\arguments{
\item{user_table_csv_paths}{list of paths, provided by the user}

\item{table_names}{list of user table names}

\item{json_template_path}{path of json file for schema template}

\item{delimiter}{csv file delimiter, default ","}
}
\value{
list(table_format_results, combined_check_df) where table_format_results 
is a data frame with format information related with user tables, and 
combined_check_df is a data frame with table and related table information
}
\description{
Table content check based on json schema template for uploaded user csv tables
}
