% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_relation_json.R
\name{extract_relation_json}
\alias{extract_relation_json}
\title{Relation table extraction from json template for a given table name}
\usage{
extract_relation_json(json_data, table_name)
}
\arguments{
\item{json_data}{list, nested json data importon from json schema file}

\item{table_name}{string, table name to be search for in the schema}
}
\value{
relation_df, a data frame if the table_name is found, 
or string "Failed" otherwise
}
\description{
Relation table extraction from json template for a given table name
}
