% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_columns.R
\name{extract_table_columns}
\alias{extract_table_columns}
\title{Extraction of list of column names for a table, from json schema}
\usage{
extract_table_columns(json_data, table_name)
}
\arguments{
\item{json_data}{list, nested json data importon from json schema file}

\item{table_name}{string, table name to be search for in the schema}
}
\value{
table_info, list of column names if the table_name is found, 
or string "Failed" otherwise
}
\description{
Extraction of list of column names for a table, from json schema
}
