% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebridgechecker_app.R
\name{rwebridgechecker_app}
\alias{rwebridgechecker_app}
\title{rwebridgechecker shiny app}
\usage{
rwebridgechecker_app(...)
}
\arguments{
\item{...}{no argument is supplied}
}
\value{
Opens the shiny app in a new window or in browser
}
\description{
In the console, type: rwebridgechecker_app()
}
